refactor(tui): move the cost scope key into the renderer
The `bitrouter/costScope` `_meta` key was split across the crate boundary:
the wire spelling and `from_usage` lived in `apps/bitrouter/src/chat/cost.rs`
while the figure and its `Scope` lived in `bitrouter-tui`, joined by a
`CostReader` function pointer the caller had to supply. The split existed to
keep one rule true — that no BitRouter name appears in the renderer — and it
bought two places to look for one key.

That rule is now retired, so the key moves in and `CostReader` goes with it.
`View::open` takes only the route; the footer calls `cost::from_usage`
directly.

The charter it replaces

`bitrouter-tui` was ACP-generic: forbidden from naming any BitRouter concept,
checked by `grep -rn "bitrouter/" crates/bitrouter-tui/src` returning nothing.
Two things were wrong with that as a mechanism. The grep never passed —
unanchored, it matched the doc comment stating it. And genericity was never
the point; it was a proxy for the property that actually matters, which is
that the renderer cannot *reach* the application. Cargo still enforces that,
and enforces it exactly as well as before.

What is kept is every honesty rule the charter was standing in for, now
pinned by named tests rather than by an absent dependency:

- an unscoped cost renders `unreported`, never `$0.00`
- an agent with no `providers/*` gets no picker, not a dead one
- a cancelled permission never resolves to consent

A non-BitRouter agent still renders correctly. It simply lands in the
honest-default branch of each of those.

Also: the chat reach-guard covers `session.rs`

It scanned `cost.rs` and `input.rs`; with the first gone it would have been
left watching 145 lines. `session.rs` passes it on substance but tripped it
on prose — its module doc named `crate::daemon` and `RouteControl` to say it
uses neither. Reworded, so the guard now covers 850 of the ~950 app-side chat
lines instead of 145.
