+++ crates/bitrouter-tui/src/lib.rs
@@ -1,51 +1,55 @@
-//! Terminal renderer for one ACP agent session.
+//! Terminal renderer for BitRouter's ACP chat session.
 //!
-//! # The boundary
+//! # The boundary, and what it is now for
 //!
-//! **This crate must never depend on `bitrouter`, the application.** That
-//! absence is the point, and it is enforced by the build rather than by
-//! prose: everything rendered here arrives over the ACP wire, and anything
-//! that cannot be learned that way is passed in by the caller.
+//! **This crate must never depend on `bitrouter`, the application.** Cargo
+//! enforces it — the app depends on this crate by path, so the reverse edge is
+//! a cyclic-package error rather than a matter of review.
 //!
-//! The previous terminal UI died of accretion — it lived inside the
+//! The previous terminal UI died of accretion. It lived inside the
 //! application, so it could reach any function in it, and it grew verbs with
-//! no command-line equivalent until nobody could say what it was. A module
-//! boundary is a promise; a crate boundary is a compiler error. This is the
-//! same rule, made checkable.
+//! no command-line equivalent until nobody could say what it was. That is
+//! still the failure this boundary prevents: what it stops is *reachability*,
+//! not vocabulary.
 //!
-//! Two consequences follow, and both are deliberate:
+//! # What this crate stopped claiming
 //!
-//! - **Session-scoped only.** There is no metering store here, no daemon
-//!   control socket, no request history — nothing daemon-wide is reachable,
-//!   so nothing daemon-wide can be drawn. What one session did is the whole
-//!   subject.
-//! - **Generality is a consequence, not a goal.** Conforming to ACP is what
-//!   makes this renderer work against agents other than BitRouter; it is not
-//!   a feature paid for separately. Anything the protocol does not carry is
-//!   the caller's to supply, not this crate's to assume.
+//! It was once ACP-generic: no BitRouter name appeared here, so it would
+//! render any conforming agent. That charter is **retired**. This is
+//! BitRouter's TUI, and it may name BitRouter's concepts — the
+//! [`cost::COST_SCOPE_META_KEY`] wire spelling is the first, and it lives here
+//! because splitting one key across a crate boundary bought nothing but two
+//! places to look.
+//!
+//! What is *not* retired is the reason genericity was chosen in the first
+//! place. Conforming to ACP is still how this renderer works at all, and a
+//! non-BitRouter agent still renders correctly — it simply lands in the
+//! honest-default branch of everything below.
 //!
 //! # Honesty
 //!
 //! A control that does not do what it appears to do is worse than an absent
-//! one. A figure whose scope is unknown is worse than a blank.
+//! one. A figure whose scope is unknown is worse than a blank. These are the
+//! rules that survived the charter, and they are pinned by tests rather than
+//! by an absent dependency:
 //!
-//! Where a control's honesty depends on something the protocol does not carry,
-//! that something is a **parameter**, not an inference — and there is no
-//! constructor that skips it. [`picker::Picker::open`] takes whether the agent
-//! serves `providers/*` at all; [`cost::Cost::new`] takes whose spend the
-//! figure is. The caller answers, because the caller is what knows; this crate
-//! makes not answering impossible rather than merely discouraged.
+//! - A cost with no scope renders as *unreported*, never as `$0.00`
+//!   ([`cost::from_usage`] returns `None`, and it is the only way in).
+//! - An agent that does not serve `providers/*` gets no picker at all, not a
+//!   dead one ([`picker::Picker::open`] returns `None`).
+//! - Cancelling a permission prompt can never resolve to consent.
 //!
-//! # The line the boundary is drawn on
+//! Where honesty depends on something the protocol does not carry, that
+//! something is still a **parameter** and not an inference: there is no
+//! constructor that skips the question. [`cost::Cost::new`] takes the scope;
+//! [`picker::Picker::open`] takes whether the surface exists.
 //!
-//! *Knowledge*, not medium. Rendering something is never by itself a reason to
-//! move it out; naming something that is not in the protocol always is. So
-//! `providers/list` renders here — [`ProviderInfo`] is an ACP schema type — but
-//! the `_meta` key BitRouter invents to carry cost scope stays in the app, and
-//! `grep -rn "bitrouter/" crates/bitrouter-tui/src` returning nothing is how
-//! that is checked.
+//! # What is still out of scope
 //!
-//! [`ProviderInfo`]: agent_client_protocol_schema::v1::ProviderInfo
+//! **Session-scoped only.** There is no metering store here, no daemon control
+//! socket, no request history — nothing daemon-wide is reachable, so nothing
+//! daemon-wide can be drawn. What one session did is the whole subject, and
+//! `bitrouter status --requests` is where the other question is answered.
 
 pub mod cost;
 pub mod editor;
